% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_bixi.R
\name{bixikwargs}
\alias{bixikwargs}
\title{Classification from individually specified features}
\usage{
bixikwargs(start_date = "2017-04-15 00:48", start_station_code = 6079,
  is_member = 1)
}
\description{
Classification from individually specified features
}
